backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
pillow==12.1.1
sqlparse==0.5.5
djangorestframework-simplejwt==5.3.1
django-filter==24.2
